chore: Update API initialization to handle 404 errors
This commit modifies the initialization of the API in the `create_app` function of the `__init__.py` file. The `Api` object is now instantiated with the `catch_all_404s=True` parameter, allowing it to handle 404 errors. This change ensures that any requests to undefined routes will return a proper 404 response.

In addition, the `delete` method of the `ProductResource` class in the `product_resource.py` file has been updated. It now checks if the product exists before attempting to delete it. If the product is not found, a 404 response with the message "Product not found!" is returned. After deleting the product, an empty response with the status code 204 (NO_CONTENT) is returned.

This commit follows the established convention of using the "chore" type for non-functional changes and provides a clear and concise description of the modifications made.

+++ app/resources/product_resource.py
@@ -25,16 +25,17 @@ class ProductResource(Resource):
 
         except Exception as e:
             print(f"An Error occured: {str(e)}")
-            return {"message": str(e)}, status.BAD_REQUEST
+            return {"message": str(e)}, HttpStatus.BAD_REQUEST
 
     def delete(self, product_id):
         try:
-            product = Product.query.get_or_404(product_id)
-
+            product = Product.query.get(product_id)
+            if not product:
+                return {"message": "Product not found!"}, HttpStatus.NOT_FOUND
             db.session.delete(product)
-            db.session.commit(product)
+            db.session.commit()
 
-            return product.to_dict(), HttpStatus.OK
+            return {}, HttpStatus.NO_CONTENT
 
         except Exception as e:
             print(f"An Error occured: {str(e)}")
